fix: solve the certs path issue and refactor code
