+++ package.json
@@ -11,7 +11,7 @@
   },
   "scripts": {
     "start": "NODE_ENV=production node server/app.js",
-    "test": "NODE_ENV=test nyc mocha --exit --timeout 15000 --require test/bootstrap.js test/config.test.js test/*.test.js",
+    "test": "NODE_ENV=test nyc mocha --exit --require test/bootstrap.js test/config.test.js test/*.test.js",
     "dev": "NODE_ENV=development node server/app.js",
     "lint": "eslint server/**/*.js",
     "generate-coverage": "NODE_ENV=test nyc report --reporter=text-lcov > coverage.lcov",
