--- test/testdata/passport-rp.pem
@@ -1,28 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDmfPz4FJQq95DN
-Ghyb22gDkmHRzzGeh5V6qOkSvTvUIUvUmvxhpBWaBsEYfbEfq7caXKroQCIv29oc
-nVLbbKcdzp988uer/+bsYObwFfyko5yeTZHRMBoVKcvSrOkbhTCIjS6OkxzGDivD
-9C1m/oI0a7X87XntfmLwKfVWu+X/Mk8dGrx2Ph2iYSahrLKMJh8SIysHYJEVq8iS
-PMRWUVkeqd55GWz+Jbj6dYMjuBrdXEYEa4o4cvA4WCPIXvIAAQtrDf7thHdSEwH6
-OZjJyw+CdrjK6JyQWr40TPrhjGEJZFKJn9VbVXE5Bl8t1lS0fzl2KEyGXGcaHbdY
-FvIwSy87AgMBAAECggEAPgXJPuquumc3EHIj267tbPShnQdfx1Sd3U2Btpi0dk9l
-a/cuzgmNBI8nu/LqzGZ1VcM0/F1xtaY4KbYOG1zAFa3JziJBRFAO8HHdAyMvgpBC
-b3jZMcz7X9GI9w6vhcfP+Rbgvd4JgqRCk67r3sjC+d84MDEpO8bkUdNAoU/bwtFg
-o8GgAN9bQrROrZGHlQVn391UdK+HeOpukj3T0/gjuugGeIOWR29BlwfY0ITuhE9d
-rIU0/ViaiHZ7gWWQBoP5ABGI1TSB3ttgjBIsnWfLQ2soCemBuhMyEcGGCM7IAELj
-H+r7l8P/CJNJ4W4/2F7QpzEQieV9rs5BSuD78fHbpQKBgQD4cTGP6puv/8AgN26t
-PyaPsTr+8L/ut6zoZZIVu31REC/B6K9qhmcSC79s53kLu/uGQnx9tPmzklFC8EFO
-+6vqhecXVl86H/0p9WUyxPkl2IXWdc6P73OQZd7NXNhGXM9nOnchwuBVjNpM0Rpg
-QBzh10rF2DXW2PxW/l/tQ/pIxQKBgQDtf/lKWr0g+1JkhJYLaym+3x1l/GWfqv+A
-BE5UQlLXhWPJmKprwNVKKaXaTqgRju7mIB3QBh4oLCnRtnzxoM1qov52yfvCCtoN
-l3CY3E6ArCnAePLj0ocYImmQkI8EpS4aTuONuB+JxPnCYdeNq83429+oGNnyK0Fi
-E3AnYNcX/wKBgQDaOaNNRMhoOf+qzMYcy2G1yfmGQjZ1G0V4BO/iwwN6lvs8GoUq
-zs/uW/9TS6dZlU+ESRZM3RCQmS9j2uGi3RHPOLsaFrYiukKH/pNbmwU2JxC4plZA
-vBYIelLQlIMsmw1bIVHPh/vBaa2+19WRQf+cqJ13V0lDZOmSCHJ+WL2Z+QKBgDeP
-zN5cRArHQQtLCV86ftWslNAbMBsegg+M8lY7/e11w1aSZKw4vK1Q+QANjrws2C72
-LINSKxPy4QWyz6AnombkIuvfbQqER2cOmsToJWVs4YepxAkQY5J92nx7lTnc+uAk
-vI+XF6m/bz7ObQMzwkgKCFTnG3XQMMelDIPp3+g5AoGBAJovuN/aFNAby4s7durj
-QUPKOTjkNMrUomxEczUheRK06WVh9NfHDaQHVXa+wugDCeyw7Vmmf0/85KZJeBiC
-JOOQ2WZVO2/7WKDu5HZ6G/xyEr+C+uHAXma12kK3q5Ul2YKOOdfBpG3MHcekBgyB
-b8cGQWiP+SB+pDP/ZLDhUCsW
------END RSA PRIVATE KEY-----
